@@ -22,25 +22,25 @@
 	"license": "ISC",
 	"homepage": "https://bitbucket.org/alirizadikici/com.epmvirtual.node#readme",
 	"dependencies": {
-		"firebase": "^5.8.2",
-		"handlebars": "^4.1.0",
-		"lodash": "^4.17.11",
+		"firebase": "^6.4.0",
+		"handlebars": "^4.1.2",
+		"lodash": "^4.17.15",
 		"node-pty": "^0.8.1",
-		"rxfire": "^3.3.2",
-		"rxjs": "^6.4.0",
-		"systeminformation": "^4.0.10",
+		"rxfire": "^3.6.11",
+		"rxjs": "^6.5.2",
+		"systeminformation": "^4.14.4",
 		"uuid": "^3.3.2",
 		"uuid-time": "^1.0.0",
-		"wrtc": "^0.3.5"
+		"wrtc": "^0.4.1"
 	},
 	"devDependencies": {
-		"@types/handlebars": "^4.0.40",
-		"@types/lodash": "^4.14.120",
-		"@types/node": "^10.12.24",
+		"@types/handlebars": "^4.1.0",
+		"@types/lodash": "^4.14.137",
+		"@types/node": "^10.14.15",
 		"@types/systeminformation": "^3.23.1",
-		"@types/uuid": "^3.4.4",
-		"ts-node": "^8.0.2",
-		"tslint": "^5.12.1",
-		"typescript": "^3.3.3"
+		"@types/uuid": "^3.4.5",
+		"ts-node": "^8.3.0",
+		"tslint": "^5.18.0",
+		"typescript": "^3.5.3"
 	}
 }
